+++ lib/ui/page/call/participant/view.dart
@@ -29,7 +29,7 @@ import '/ui/widget/member_tile.dart';
 import '/ui/widget/modal_popup.dart';
 import '/ui/widget/outlined_rounded_button.dart';
 import '/ui/widget/progress_indicator.dart';
-import '/util/navigator.dart';
+import '/util/platform_utils.dart';
 import 'controller.dart';
 
 /// [OngoingCall.members] enumeration and administration view.
