+++ lib/ui/page/call/screen_share/view.dart
@@ -26,7 +26,7 @@ import '/ui/page/call/widget/video_view.dart';
 import '/ui/widget/modal_popup.dart';
 import '/ui/widget/outlined_rounded_button.dart';
 import '/ui/widget/progress_indicator.dart';
-import '/util/navigator.dart';
+import '/util/platform_utils.dart';
 import 'controller.dart';
 
 /// View for selecting display for screen sharing.
