+++ lib/ui/page/home/page/chat/forward/controller.dart
@@ -47,7 +47,7 @@ class ChatForwardController extends GetxController {
     this.text,
     this.pop,
     this.attachments = const [],
-    this.onSubmitted,
+    this.onSent,
     required this.from,
     required this.quotes,
   });
@@ -74,8 +74,8 @@ class ChatForwardController extends GetxController {
   /// [Attachment]s to attach to the [quotes].
   final List<Attachment> attachments;
 
-  /// Callback, called after the [send] was submitted.
-  final void Function()? onSubmitted;
+  /// Callback, called when the [quotes] are sent.
+  final void Function()? onSent;
 
   /// Indicator whether there is an ongoing drag-n-drop at the moment.
   final RxBool isDraggingFiles = RxBool(false);
@@ -173,7 +173,7 @@ class ChatForwardController extends GetxController {
 
           await Future.wait(futures);
           pop?.call();
-          onSubmitted?.call();
+          onSent?.call();
         } on ForwardChatItemsException catch (e) {
           MessagePopup.error(e);
         } catch (e) {
