+++ lib/ui/page/home/page/chat/info/add_member/view.dart
@@ -23,7 +23,7 @@ import '/l10n/l10n.dart';
 import '/ui/page/call/search/controller.dart';
 import '/ui/widget/modal_popup.dart';
 import '/ui/widget/progress_indicator.dart';
-import '/util/navigator.dart';
+import '/util/platform_utils.dart';
 import 'controller.dart';
 
 /// [Chat.members] addition view.
