+++ lib/ui/page/home/page/my_profile/blacklist/view.dart
@@ -25,7 +25,7 @@ import '/themes.dart';
 import '/ui/page/home/widget/contact_tile.dart';
 import '/ui/widget/modal_popup.dart';
 import '/ui/widget/widget_button.dart';
-import '/util/navigator.dart';
+import '/util/platform_utils.dart';
 import 'controller.dart';
 
 /// View displaying the blacklisted [User]s.
